+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove `Input::new_unchecked` and `Input::next`.
 - Remove implementation of `Action` for `&mut Action`.
+- Remove `Eat<&String>`. Use `Eat<&str>` and `eat(s.as_str())` instead.
 - Remove `eat_unchecked`. Use `wrap` instead.
 
 ## [0.4.0] - 2025-01-10